% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bijarnia_mahay_2022}
\alias{bijarnia_mahay_2022}
\title{SMPD4-Related Clinical Phenotype Data Set from Bijarnia-Mahay et al. (2022)}
\format{
A tibble: 1 x 152

\describe{
\item{_ind}{columns ending in _ind are indicator columns created by merging duplicate tokenised terms into one feature.}
}
}
\source{
\url{https://doi.org/10.1038\%2Fs10038-021-00981-3}
}
\usage{
bijarnia_mahay_2022
}
\description{
A dataset of clical features and phenotypes of Neurodevelopmental Disorder with Arthrogryposis, Microcephaly and Structural Brain Anomalies Caused by Bi-allelic Partial Deletion of SMPD4 gene.
}
\details{
This data has been transformed into an analysis-ready tidy data set from Bijarnia-Mahay, S., Somashekar, P. H., Kaur, P., Kulshrestha, S., Ramprasad, V. L., Murugan, S., ... & Shukla, A. (2022). Growth and neurodevelopmental disorder with arthrogryposis, microcephaly and structural brain anomalies caused by Bi-allelic partial deletion of SMPD4 gene. Journal of human genetics, 67(3), 133-136.
}
\keyword{datasets}
