% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CADD}
\alias{CADD}
\title{Variant and CADD data}
\format{
A tibble: 20 x 3

\describe{
\item{\code{family}}{Patient family identifier from study}
\item{\code{variant_type}}{Type of SNV}
\item{\code{cadd}}{Combined Annoation Dependent Depletion CADD score where available}
}
}
\usage{
CADD
}
\description{
Combined Annoation Dependent Depletion (CADD) data and Single Nucleotide Variant
type aggregated from studies on SMPD4 related variants.
}
\details{
Data are manually aggregated from:

Bijarnia-Mahay, S., Somashekar, P. H., Kaur, P., Kulshrestha, S., Ramprasad, V. L., Murugan, S., ... & Shukla, A. (2022). Growth and neurodevelopmental disorder with arthrogryposis, microcephaly and structural brain anomalies caused by Bi-allelic partial deletion of SMPD4 gene. Journal of human genetics, 67(3), 133-136.

Magini, P., Smits, D. J., Vandervore, L., Schot, R., Columbaro, M., Kasteleijn, E., ... & Mancini, G. M. (2019). Loss of SMPD4 causes a developmental disorder characterized by microcephaly and congenital arthrogryposis. The American Journal of Human Genetics, 105(4), 689-705.

Ji, W., Kong, X., Yin, H., Xu, J., & Wang, X. (2022). Case Report: Novel Biallelic Null Variants of SMPD4 Confirm Its Involvement in Neurodevelopmental Disorder With Microcephaly, Arthrogryposis, and Structural Brain Anomalies. Frontiers in Genetics, 13.

Ravenscroft, G., Clayton, J. S., Faiz, F., Sivadorai, P., Milnes, D., Cincotta, R., ... & Davis, M. R. (2021). Neurogenetic fetal akinesia and arthrogryposis: genetics, expanding genotype-phenotypes and functional genomics. Journal of medical genetics, 58(9), 609-618.

Monies, D., Abouelhoda, M., Assoum, M., Moghrabi, N., Rafiullah, R., Almontashiri, N., ... & Alkuraya, F. S. (2019). Lessons learned from large-scale, first-tier clinical exome sequencing in a highly consanguineous population. The American Journal of Human Genetics, 104(6), 1182-1201.
}
\keyword{datasets}
