% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ji_2022}
\alias{ji_2022}
\title{SMPD4-Related Clinical Phenotype Data Set from Ji et al. (2022)}
\format{
A tibble: 1 x 152

\describe{
\item{_ind}{columns ending in _ind are indicator columns created by merging duplicate tokenised terms into one feature.}
}
}
\source{
\url{https://doi.org/10.3389\%2Ffgene.2022.872264}
}
\usage{
ji_2022
}
\description{
A dataset of clical features and phenotypes of Neurodevelopmental Disorder with Arthrogryposis, Microcephaly and Structural Brain Anomalies Caused by Bi-allelic Partial Deletion of SMPD4 gene.
}
\details{
This data has been transformed into an analysis-ready tidy data set from Ji, W., Kong, X., Yin, H., Xu, J., & Wang, X. (2022). Case Report: Novel Biallelic Null Variants of SMPD4 Confirm Its Involvement in Neurodevelopmental Disorder With Microcephaly, Arthrogryposis, and Structural Brain Anomalies. Frontiers in Genetics, 13.
}
\keyword{datasets}
