% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{magini_2019}
\alias{magini_2019}
\title{SMPD4-Related Clinical Phenotype Data Set from Magini et al.(2019)}
\format{
A tibble: 21 x 152

\describe{
\item{_ind}{columns ending in _ind are indicator columns created by merging duplicate tokenised terms into one feature.}
}
}
\source{
\url{https://doi.org/10.1016/j.ajhg.2019.08.006}
}
\usage{
magini_2019
}
\description{
A dataset of clical features and phenotypes of Neurodevelopmental Disorder with Arthrogryposis, Microcephaly and Structural Brain Anomalies Caused by Bi-allelic Partial Deletion of SMPD4 gene.
}
\details{
This data has been transformed into an analysis-ready tidy data set from Table S2 of Magini, Pamela, et al. "Loss of SMPD4 causes a
developmental disorder characterized by microcephaly and congenital arthrogryposis." The American Journal of Human
Genetics 105.4 (2019): 689-705.
}
\keyword{datasets}
