% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monies_2019}
\alias{monies_2019}
\title{SMPD4-Related Clinical Phenotype Data Set from Monies et al. (2019)}
\format{
A tibble: 2 x 152

\describe{
\item{_ind}{columns ending in _ind are indicator columns created by merging duplicate tokenised terms into one feature.}
}
}
\source{
\url{https://doi.org/10.1016\%2Fj.ajhg.2019.04.011}
}
\usage{
monies_2019
}
\description{
A dataset of clical features and phenotypes of Neurodevelopmental Disorder with Arthrogryposis, Microcephaly and Structural Brain Anomalies Caused by Bi-allelic Partial Deletion of SMPD4 gene.
}
\details{
This data has been transformed into an analysis-ready tidy data set from Monies, D., Abouelhoda, M., Assoum, M., Moghrabi, N., Rafiullah, R., Almontashiri, N., ... & Alkuraya, F. S. (2019). Lessons learned from large-scale, first-tier clinical exome sequencing in a highly consanguineous population. The American Journal of Human Genetics, 104(6), 1182-1201.
}
\keyword{datasets}
