% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ravenscroft_2020}
\alias{ravenscroft_2020}
\title{SMPD4-Related Clinical Phenotype Data Set from Ravenscroft et al. (2020)}
\format{
A tibble: 3 x 152

\describe{
\item{_ind}{columns ending in _ind are indicator columns created by merging duplicate tokenised terms into one feature.}
}
}
\source{
\url{https://doi.org/10.1136\%2Fjmedgenet-2020-106901}
}
\usage{
ravenscroft_2020
}
\description{
A dataset of clical features and phenotypes of Neurodevelopmental Disorder with Arthrogryposis, Microcephaly and Structural Brain Anomalies Caused by Bi-allelic Partial Deletion of SMPD4 gene.
}
\details{
This data has been transformed into an analysis-ready tidy data set from Ravenscroft, G., Clayton, J. S., Faiz, F., Sivadorai, P., Milnes, D., Cincotta, R., ... & Davis, M. R. (2021). Neurogenetic fetal akinesia and arthrogryposis: genetics, expanding genotype-phenotypes and functional genomics. Journal of medical genetics, 58(9), 609-618.
}
\keyword{datasets}
