% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smpd4_phenotype}
\alias{smpd4_phenotype}
\title{Summary of SMPD4-Related Clinical Phenotype}
\format{
A tibble: 21 x 152
}
\source{
\url{https://doi.org/10.1016/j.ajhg.2019.08.006}
}
\usage{
smpd4_phenotype
}
\description{
A dataset from Table S2 of Magini, Pamela, et al. "Loss of SMPD4 causes a
developmental disorder characterized by microcephaly and congenital
arthrogryposis." The American Journal of Human
Genetics 105.4 (2019): 689-705.
}
\details{
This data has been transformed into an analysis-ready tidy data set.
}
\keyword{datasets}
